Fixed minor assert() bug

+++ code/common/shared.cpp
@@ -2,7 +2,7 @@
 #include "shared.h"
 
 ui8 SafeCastIntToUI8(int Value) {
-    Assert(Value <= INT_MAX);
+    Assert(Value <= UINT8_MAX);
     ui8 Result = (ui8)Value;
     return(Result);
 }
